[dg] Support builds with uv workspace (#33088)
## Summary & Motivation

Two changes to how `dg plus deploy` works:

- The default Dockerfiles now build the `uv` env with a `--package
<PACKAGE_NAME>` arg. This is redundant when the `dg` project is outside
of a uv workspace, but it is necessary when the package is inside a
workspace.
- The build context directory when deploying will be set to the uv
workspace root if we're in a workspace, otherwise to the project root
(current behavior). Setting to the uv workspace root allows workspace
packages to resolve during the build.

A caveat is that a lot more might get copied into the image since we
basically add the full content of the build context into the image (`ADD
. /app` during build)-- but I think that can be refined later.

## How I Tested These Changes

Manually + existing test suite.