Don't include descriptions and metadatas on job snapshots for outputs of asset job ops (#33092)
### Summary

For assets, these descriptions and metadata can already be found on the asset graph, and can be quite large. Including them on the output metadata in the snapshot as well (especially for __ASSET_JOB and subsets therein) significantly bloats the size of the snapshot.

On a representative large snapshot of a single dbt node, byte size is reduced by about 10x: (and the metadata is of course still available on the asset graph)

```
Original size: 52,271,581 bytes
Stripped size: 4,774,884 bytes
Size reduction: 47,496,697 bytes (90.87%)
```

### Test Plan

new test cases, sanity checked local UI in the job snapshot page

### Changelog

Performance improvements when launching runs of asset jobs, by removing unnecessary op output metadata on the job snapshot that is computed for each run.