Allow rendering per-partition statuses for backfills for assets no longer in the graph (#31198)
## Summary & Motivation
Previously, this method on AssetGraphSubset had an optional asset_graph
argument that could be used to add some validation and fall back to the
empty subset. Now, AssetGraphSubset is more of a standard data class
that raises an error if the passed in asset key . Validation of that
sort is reserved for classes like AssetGraphView

This allows us to fix a longstanding papercut with backfills where we
couldn't fetch historical partition status for backfills that involved
assets that are no longer in the graph. Not everything is perfect there
(e.g. we can't show the 'target root partitions' still), but we can now
show basic stat information again. We use the fact that the backfills
always store the target_subset, which is a superset of all the other
subsets, to fall back to an empty subset if the key isn't part of the
materialized or failed subsets yet.

## How I Tested These Changes
BK, view a backfill that was previously erroring due to missing asset
key information, no longer is

## Changelog
Fixed an issue where removing an asset prevented rendering status
information for backfills involving that asset in the Dagster UI.